frozen_requirements.txt
##Python 3.12.9
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
branca==0.8.1
cachetools==6.1.0
catboost==1.2.8
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.1
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cvxpy==1.6.6
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
faiss-cpu==1.11.0
Faker==37.4.0
fastapi==0.115.14
filelock==3.18.0
folium==0.20.0
fonttools==4.58.4
fsspec==2025.5.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.21
greenlet==3.2.3
h11==0.16.0
haversine==2.9.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-core==0.3.71
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langgraph==0.5.4
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.45.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
openai==1.93.0
orjson==3.10.18
ormsgpack==1.10.0
osqp==1.0.4
packaging==24.2
pandas==2.3.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
PuLP==3.2.1
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
scs==3.2.7.post2
seaborn==0.13.2
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
xgboost==3.0.2
xxhash==3.5.0
xyzservices==2025.4.0
zstandard==0.23.0
